refactor: flatten gui
