--- src/cala/gui/nodes/__init__.py
@@ -1,9 +1,0 @@
-from .component_counter import ComponentCounter, ComponentCounterParams
-from .frame_streamer import FrameStreamer, FrameStreamerParams
-
-__all__ = [
-    "FrameStreamer",
-    "FrameStreamerParams",
-    "ComponentCounter",
-    "ComponentCounterParams",
-]
