+++ src/cala/nodes/gui/component_counter.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 import xarray as xr
 
 from cala.assets import Footprints
-from cala.gui.nodes.util import send_through
+from cala.gui.util import send_through
 
 
 @dataclass
