+++ src/cala/nodes/gui/frame_streamer.py
